meta-ampere: flashrom: support flashing less than chip size
flashrom just supports flashing image which has equal size of the
SPI-NOR device that the image is flashed on. However, firmware image
rarely has the same size and we might need to patch to the image to be
able to use the tool to flash the firmware. And it is not always
possible.
This commit supports firmware image which is less than the chipsize to
remove this limitation.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>