meta-ampere: state: emit dbus signals while initialization
Currently, the phosphor-host-state-manager does not emit dbus signals to
inform other applications about the initialization value.
Some applications that started before phosphor-host-state-manager can
not detect the host's state if they are monitoring dbus signals.

Example: Sometimes, phosphor-host-state-manager fails to start in the
first time, and it has to restart. During that, Dbus-sensors's
applications started and monitored dbus signals to check the host's
state. But they never receive dbus signals when
phosphor-host-state-manager sets the initialization value. If the
initialization value of the host's state is "Running", it made sensors
that depend on the host's state can not be probed, their "Available"
property is always "false".

This commit removes "skipSignal" flags when phosphor-host-state-manager
updates the initialization values.

Tested:
    1. Turn on the host, then reboot BMC.
    2. Force to start phosphor-host-state-manager after dbus-sensors
       applications.
    3. Check sensors
       $ipmitool sdr list.
    4. Sensors that depend on the host's state are shown correctly.

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>