meta-ampere: mtjefferson: support Host Secure Provision procedure
BMC shall assert SPECIAL_BOOT GPIO to the CPU to trigger Host Secure
Provision. Then BMC shall reset the system, or turn on the system if
it is being OFF and shows the message of success or failure to bmc
console.
After the host enters Provisioning Boot Mode successfully (or having
failure), BMC shall deassert the SPECIAL_BOOT GPIO to LOW.

Tested:
1. Flash FW SECPro and enter the Host Secure Provisioning Boot Mode
   $ ampere_flash_bios.sh spinor.img 1 1
   BMC console will show:
   ======================================================================
   =                                                                    =
   =   SECPROV IS SUCCESSFUL, PLEASE RE-INSTALL THE NORMAL BOOT IMAGE   =
   =                                                                    =
   ======================================================================
2. The Host console will show:
   [00:00:00.145,000] <inf> SECpro_cert_ext: assert FW_BOOT_OK
   [00:00:00.145,000] <inf> SECpro_cert_ext: deassert FAULT_ALERT
   [00:00:00.145,000] <inf> SECpro_cert_ext: SEC provision completed successfully

Signed-off-by: Hieu Huynh <hieuh@os.amperecomputing.com>