meta-ampere: mtjefferson: fan: Handle sensor missing behavior
This commit adds a local patch to phosphor-pid-control to update the
behaviors in handling the configuration of MissingIsAcceptable. More
details can be found in the comment section in the patch. This commit
allows important sensors to send the fan zone for failsafe mode when
they are not available on DBus.

Tested with PLDM sensors:
BMC boots up, PLDM and phosphor-pid-control services are running
normally => fans are controlled following the table, supposing that all
other thermal sensors are functioning normally.

1. Stop PLDM service => Fans are set to 100% PWM
2. Start PLDM service => Fans are controlled following the table after a
   while.

Use {BPN} in the recipe. Refer to https://docs.yoctoproject.org/ref-manual/qa-checks.html

Signed-off-by: Chau Ly <chaul@amperecomputing.com>