meta-ampere: webui: add user management patches
This fixes the following issues:


## Cannot unlock users

The main cause is that it gets stuck when calling the password
dependency. However, the exact reason for this is unknown. Therefore,
to fix this issue, I have written a new function to handle it.
Opened bug for the community at [1]

Tested:
 1. Log in to BMC Web and go to the user-management.
 2. Adding a new user is successful.
 3. Editing a user is successful.

[1]. https://github.com/openbmc/webui-vue/issues/129


## Cannot unlock users

Currently, in manual mode, when a user is locked, the edit user page
does not allow unlocking the user. This commit enables unlocking
locked users manually via the edit user page.

Tested:
 - Exceeded allowed login attempts to lock the user.
 - Used root account to access edit user and unlocked the user successfully.


Signed-off-by: Huy Anh Le <hule@amperecomputing.com>