meta-ampere: flash: update the method to turn on the Host after flashing
When BMC is flashing firmware, the ActivateBlockTransiton interface is
added. The PSM checks the ActivateBlockTransiton exits or not, if this
interface exits then Transition request is rejected. Otherwise, we handles
transition requests. Therefore, we have to check the ActivateBlockTransiton
interface before requesting to turn the Host on after flashing.

Tested:
    1. Turn the Host on.
    2. Request to flash new firmware via Redfish/WebUi.
    3. During flashing, Host is turned off.
    4. When flashing done. Host is turned on.

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>