meta-ampere: mtjefferson: Add driver binder when power on HOST
The PCIe NIC card is off once the HOST is powered off. That leads to
the PCIe NIC temperature sensor is probed unsuccessfully when Host
switches from power-off state to power-on state.

The driver binder script will help the PCIe NIC temperature sensor
is probed again after Host is powered on.

Tested:
1. AC cycle power
2. Power ON HOST
3. Check the PCIeNIC_MaxTemp sensor by "ipmitool sdr list".

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>