meta-ampere: mtjefferson: Retry driver binder when failure
Some of the hardware devices are not available after the P12V_MAIN_PG
signal was triggered, because of latency hardware.

For example on the PCIe NIC card - ConnectX-4 LX SFP28 MCX4121A-ACAT -
Firmware version: 14.32.1010, the P12V_MAIN_PG signal was triggered,
but the temperature sensor on the PCIe NIC card is not available.
Need to retry driver binder if the binder is failed.

Config to the systemd to restart service 5 times with interval time is
2s.

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>