meta-ampere: ipmid: correct type code of Power sensors
When users request to get the sensor properties via "ipmitool
sensor get <sensor_name>" command, the "Sensor Type" of Power sensors
are "Other".

Root cause: In the dynamic-sensor mode, the "Sensor Type" of a sensor is
indicated via the object path. Currently, the if the object path
includes "/power/" inside, the "Sensor Type" is "Other".

Solution: Update the table to map "power" to "Power Supply" type.

Tested:
   1. Get the power sensor via ipmitool commands
      ipmitool sensor get <power_sensor_name>
   2. The Sensor Type is "Power Supply"

Signed-off-by: Thang Tran <thuutran@amperecomputing.com>