meta-ampere: usbnet: reconfigure device after setting new properties
There is a chance that after usb-ctrl, it takes too long for our
ampere_usbnet.sh script to reconfigure the properties. When that
happens, new properties will not get applied.

Reproduce steps:

1. Simulate a one second delay, like below

        ...
        usb-ctrl ecm usbnet on ...
        delay 1
        echo 0x0103 > /sys/...
        echo "OpenBMC usbnet Device" > /sys/...
        ...

2. On BMC, run ampere_usbnet.sh
3. On Host, run lsusb -v | grep iProduct
   Expect: no OpenBMC usbnet Device, only OpenBMC USB Device.

The fix is to re-echo into UDC to reconfigure the USB gadget. This trick
is used inside usb-ctrl source itself.

Tested: Even with the 1 second delay above, the USB Gadget is created
with the correct name.

Signed-off-by: Khang D Nguyen <khangng@os.amperecomputing.com>