meta-ampere: dump-manager: Add serialization support for elapsed attribute
Currently, the elapsed time of faultlog entry is lost after the dump-manager
service restart or bmc reboot. It makes the created time of faultlog entry
show in redfish changed.

The implemented serialization of dump entry attributes is now only supported
for originatorId, originatorType, and startTime. We need add support for
the elapsed attribute.

Tested:
1. Trigger bert: ipmitool raw 0x3c 0x15
2. Check Created time of cper/bert entries
3. Reboot BMC or restart dump-manager
4. Make sure Created time of cper/bert entries are unchanged.
5. Reboot BMC or restart dump-manager
4. Make sure cper/bert entries has all attributes are unchanged.

Signed-off-by: Dung Cao <dung@os.amperecomputing.com>