WORKAROUND: meta-ampere: bmcweb: Add ipv4StaticAddresses checking
Follow code flow from [1], When we patch a redfish command
to update IPv4StaticAddresses then the handleDHCPPatch also
is called before.

That makes the defaultGateway is removed, because there is
no any IPv4 static address existed.

This commit will add IPv4StaticAddresses command checking
before remove the defaultGateway. If the IPv4StaticAddresses
command is being called, then the defaultGateway will be not
removed.

[1] https://github.com/openbmc/bmcweb/blob/master/redfish-core/lib/ethernet.hpp#L2416

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>