meta-ampere: bmcweb: Remove ipv4GwUpdated checking when update default gateway
In the async getProperty thread, The ipv4GwUpdated is used to check
if IPv4 default gateway was added. The ipv4GwUpdated is defined as
a local variable. It will be released resource once the getSubTree’s
lambda exits.

That may make the ipv4GwUpdated value is not correct once the
getSubTree’s lambda exits before the getProperty thread completes.

Follow the Redfish and IPMI, the user just can add only one IPv4 address
on the ethernet interface. So, the ipv4GwUpdated checking is not necessary.

This commit will remove the ipv4GwUpdated variable to prevent
the potential issues.

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>