meta-ampere: Refactor Ampere fan control
This commit updates the ampere_fanctrl.sh script on Ampere's platforms to
not allow enabling fan control when it's already running, or disabling when
it's already off. It will not return an error to callers in those cases, but
will prevent unnecessary actions to systemd. systemd masking/unmasking
service scheme will be replaced by utilizing another service to conflict
with phosphor-fan services and prevent them from starting when this service
is active. The new service is named "ampere-fanctrl-manual.service" and is
added to systemd under ampere-utils recipe. Correspondingly, phosphor-fan
services (andphosphor-pid-control service in Mt.Jefferson) need an extra
ExecCondition to check active state of "ampere-fanctrl-manual.service" prior
to starting. These services won't exit with error code when
"ampere-fanctrl-manual.service" is active, but just skip all later
executions. "ampere-fanctrl-manual.service" is started/stopped following
the IPMI OEM command to disable/enable fan control.

Moreover, this commit also reformats the script's syntax to bypass shellcheck
error SC2004 and define some variables for structural purpose.

For Mt.Jefferson, fan control status will not depend on
phosphor-pid-control.service if host power status is Off.

With this change, when BMC can not respond in-time to IPMI OEM command of
enabling/disabling fan control and the client retries the request, this
script will not execute the actions multiple times and impact next commands.

Tested:

Test case 1:

BMC is on, Host is on.

1. ipmitool -H <IP> -U root -P <PASS> -C 17 -I lanplus raw 0x3c 0x03
   0x01 // to disable fan control
2. ipmitool -H <IP> -U root -P <PASS> -C 17 -I lanplus raw 0x3c 0x02

Expect no error log when the second command is executed. Same to the
case of enabling fan control.

Test case 2:

BMC is on, Host is on.

1. Disable fan control with "ipmitool raw 0x3c 0x03 0x01"
2. Turn off host power => fans preserve the previous speeds
3. Turn on host power => phosphor-fan services are not started
4. Enable fan control with "ipmitool raw 0x3c 0x03 0x00" => phosphor-fan
   services are started

Test case 3 (for Mt.Jefferson):

BMC is on, Host is on.

1. Turn off host power => fans run at 10% PWM
2. Disable fan control => fans preserve the previous speeds (10%)
3. Enable fan control => phosphor-fan services are started, but
   phosphor-pid-control service is not started, fans still run at 10%.
4. Get fan control status => Enabled (when host is Off, does not care
   about phosphor-pid-control)
5. Turn on host power => all fan services starts

Signed-off-by: Chau Ly <chaul@amperecomputing.com>