meta-ampere: fix unexpected sel of GPIO System Power Good when AC/DC
Unexpected sel of power good when AC/DC system.

Root cause:
The ampere-host-on-host-check is a service to check the power on
sequence to analyze power on failure. It will check the Chassis power
good to log relevant events.
During the regression testing of AC/DC system, the GPIO system power good
is asserted before the pgood signal (power-chassis-good). Therefore event
logs related to power supply failure will be generated.

Solution:
To ensure that the chassis power state is asserted (pgood = 1) prior to
initiating the power-on sequence, the After=obmc-chassis-poweron@%i.target
has been added into the ampere-host-on-host-check.

Tested:
1. AC/DC system without event log of power supply failure
2. ipmitoo sdr can show sensor list
3. Reboot BMC, AC/DC without any error/warning logs

Signed-off-by: Hieu Huynh <hieuh@os.amperecomputing.com>