meta-ampere: user: manually lock user after exceeding allowable failed attempts
Currently, when configured in Manual mode, if a user exceeds
the allowed number of failed login attempts:
- The Locked field on Redfish/DBus is not set to true.
- The Enabled field remains true (the user is not disabled).
However, in reality, the user cannot log in (PAM backend denies access).
This causes confusion for administrators because there is no attribute
in the UI indicating that the user is manually locked.
This commit updates the logic to accurately reflect the account's locked
status when the user is manually locked due to too many failed login attempts.

Tested:
 - Verified that after exceeding allowed login attempts in Manual mode,
   the account lock status is correctly reflected.
 - Checked that the Locked field is updated on Redfish/DBus and user cannot
   log in as expected. Locked as True. Command to check:
   Dbus: busctl get-property xyz.openbmc_project.User.Manager \
		/xyz/openbmc_project/user/user1 \
		xyz.openbmc_project.User.Attributes \
		UserLockedForFailedAttempt
   Redfish: curl -X GET --user root:0penBmc --insecure \
		-H "Content-Type: application/json" \
		https://<bmc_ip>/redfish/v1/AccountService/Accounts/user1

Signed-off-by: Huy Anh Le <hule@amperecomputing.com>