meta-ampere: crash-capture: Simplify Bert dependencies of power sequence
Currently, the ampere-bert-power-handle.service adds many dependencies
on the power sequence to make sure BERT is started after host is
powered off, and finished before host is turned on.

This may affect to the power sequence and cause some bug during stress
bert or power sequence. So this commit simplifies the Bert dependencies.

Test should be done:
1. Stress test power cycle/reset
2. Stress AC
3. Stress trigger CrashDump via ipmitool

Signed-off-by: Dung Cao <dung@os.amperecomputing.com>