meta-ampere: host-reboot: make sure power is off before on
The bug happens when `obmc-chassis-powered-off@.target` is activated
after ampere_power_utils.sh transition power state to on. This causes
`obmc-powered-off@0.service` is failed to start
`obmc-chassis-powered-off@.target` when power is transitioning to on. So
`obmc-powered-off@0.service` is failed to auto deactivated.

Failed to deactivated `obmc-powered-off@0.service` then causes
`obmc-chassis-poweron\@0.target` can't be started successfully this
cause `CurrentPowerState` can't change to
"xyz.openbmc_project.State.Chassis.PowerState.On".

But some onboard sensors uses this signal to identify the power is
already on and starts updating the sensors data.

So `obmc-chassis-powered-off@.target` should be activated before starts
sequence to turn on the host. Or activating of
`obmc-chassis-powered-off@.target` can be used to identify the system is
ready to start sequence to turn on the host.

Tested:
1. Stress BERT 500 time, make sure there is no issue.
2. Stress DC 500, time make sure there is no issue.
3. Stress trigger `reboot` from the host 100 times, make sure there is
no issue.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>