meta-ampere: host-reboot: Prevent infinite wait for the power off
There should be the time out when turning off the host. Because the
power is usually turned off in less than 1 minutes after user trigger
turn off action in ipmitool, redfish or bmc web. So applying 2 minutes
timeout in turning off.
The error will be printed and the system will be keeps to off for future
investigating.

Tested:
1. Stress Bert or call `reboot` in host console to start `host-reboot`
sequence which will trigger the changing.
2. Make sure the system is powered on successfuly and no error message
`Failed to turn off the chassis power after 120s.`
3. Repeat the test 500 times.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>