meta-ampere: Disable Redfish SEL for systemd unit failures
Although dependencies are already  attached to systemd
services to ensure correct starting order and avoid conditioning
failures, some programs executed by these services require many
more conditions to be successfully started (D-Bus object paths,
system files,…), and these conditions sometimes can’t be known
prior to starting so can’t be placed in the service configuration
units. Therefore, some services are configured to always be
restarted for daemon-based services and restarted on failures for
action-based services.

As we are having a CrashTemplate in rsyslog configuration, those
failures will be parsed from journal log and log into Redfish
Event Log under the "OpenBMC.0.1.ServiceFailure" registry. This is
annoying to users as they don't need to be aware of the failure of any
systemd unit, and these failures can't help separate a permanent failure
or critical failure from failures that are followed by a restart.

This commit removes CrashTemplate configuration, and revert the trick we
did before for phosphor-fan services where we prefixed `ExecStart` with
“-” to ignore the exit codes (if exiting, any failed execution will be
considered as success).

This commit also restructures and renames some overriding configurations
to simplify phosphor-fan service configurations.

After this commit, everything below Redfish will function as usual,
except that phosphor-fan services can exit unsucessfully when some D-Bus
objects are not present and then be restarted, but systemd unit failures
will not be visible to users in Redfish Event Log.

Tested:

1. AC cycle
2. Wait until BMC state is Ready
3. GET /redfish/v1/Systems/system/LogServices/EventLog/Entries

It's expected that there's no entry with `MessageId`
"OpenBMC.0.1.ServiceFailure".

Signed-off-by: Chau Ly <chaul@amperecomputing.com>