meta-ampere: PDI: Add `BootStrapAccount` property
Add `BootStrapAccount` property to `User.Attributes` D-Bus interface to
identify if the user is bootStrap account as design [1].
Phosphor-user-manager will host `User.Attributes`, and the
phosphor-ipmi-host will update `BootStrapAccount` property for the
created bootStrap account when it handles ipmi command `Get bootstrap
account credentials (NetFn 2Ch, Command 02h)` as DSP0270 V1.3.1.
The BMCWeb socket instance handling a redfish request in the
redfish host interface will identify if the user in the request is
from bootstrap account base on this property. Then it will authenticate for
the request.

[1] https://gerrit.openbmc.org/c/openbmc/docs/+/79327

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>