meta-ampere: mtmitchell: Enable DisplayPort
To enable DisplayPort feature, the 'display_port' node must be
enabled in the U-Boot device tree, and 'CONFIG_ASPEED_DP' must be
enabled in the U-Boot configuration.

This commit adds 'CONFIG_ASPEED_DP' to Mt.Mitchell's U-Boot
configuration. Additionally, the patch enabling the 'display_port' node
in 'ast2600-ampere.dts' is moved to meta-common, which it is
applicable to both Mt.Mitchell and Mt.Jefferson platforms.

Tested:
1. The host OS console is successfully displayed on a monitor via DisplayPort.
2. After a BMC reboot, the monitor continues to display the host OS console.

Signed-off-by: Chanh Nguyen <chanh@os.amperecomputing.com>