meta-ampere: EM: Revert "Remove redundant Backplane Version"
In order to dynamically support display of Backplane version on boards
that either have Backplane Board Extra field in FRU data or not, the
"Version" property should be configured for the Backplane board entity.
Therefore, this commit reverts the patch that "Remove redundant
Backplane Version". An upstream commit [1] will set an empty string to
the property on D-Bus when Board Extra field is not present. This will
result in an empty string on Redfish query to this board also.

[1]: https://gerrit.openbmc.org/c/openbmc/entity-manager/+/82998?usp=dashboard

Tested:

curl -X GET --user root:${bmc_password} -H "Content-Type: application/json" \
--insecure https://${bmc_ip}/redfish/v1/Chassis/Jefferson_BP_1 | jq

1. With Backplane that has Board Extra field in FRU
=> "Version": <Board Extra field from FRU>

2.  With Backplane that has no Board Extra field in FRU
=> "Version": ""

Signed-off-by: Chau Ly <chaul@amperecomputing.com>