meta-ampere: webui: Update user account edit permission control based
Currently, users with Operator or ReadOnly privileges cannot change
their own password because the update request includes the RoleId
property in the request body. When this property is included, the
system returns the error: "The property RoleId is not in the list
of valid properties for the resource".
This change ensures that: Users with Operator or ReadOnly privileges
are only allowed to update their password; other fields, such as
RoleId, are excluded from their update requests. The UI disables
privilege selection if the current user is not an Administrator.

Tested:
- Logged in with an Administrator account and created accounts
  with Operator and ReadOnly privileges.
- Login with Operator and ReadOnly accounts and verify that changing
  the password is successful.

Signed-off-by: Huy Anh Le <hule@amperecomputing.com>